rename ascwds worker job files
